Rename some harmony patches
