+++ Source/Harmony/ApplicationLauncher.cs
