+++ Source/Harmony/DifficultyOptionsMenu.cs
