+++ Source/Harmony/StrategySystem.cs
